requirements.txt
asgiref==3.8.1
bleach==6.1.0
diff-match-patch==20241021
Django==5.1.2
django-import-export==4.2.0
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
pillow==11.0.0
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
webencodings==0.5.1
whitenoise==6.7.0
